+++ src/ui/tvalue.cpp
@@ -1,5 +1,5 @@
 #include <ui/tvalue.hpp>
 
 TValue operator"" _px(unsigned long long value){
-    return {PIXELS, value};
+    return {PIXELS, static_cast<int>(value)};
 }
